refactor(timeutil): 2. extract profileutil's TimeProvider into a shared package
certificateutil needs the same clock seam that ProfilePrinter uses, so a
CertificatePrinter can render validity against an injected time instead of the wall
clock. It cannot import profileutil for it: profileutil already imports
certificateutil, so that is a cycle. Duplicating the interface would leave go-xcode
with two identical declarations of one concept.

Moves the canonical declaration to a new timeutil package, which depends on nothing
but `time` and so can be imported from either side. The moved file is unchanged
apart from its package clause.

profileutil keeps TimeProvider and DefaultTimeProvider as type aliases: both are
published in v2.0.0-alpha.84, so removing them would break anyone pinned there.
Aliases, not wrappers, so an implementation written against either name satisfies
both. In-repo code uses the canonical timeutil names — profile_printer.go and the
one production caller, exportoptionsgenerator/codesign_group_provider.go:52 — and
the aliases exist only for external consumers.

Worth noting for a future cleanup: ProvisioningProfileInfoModel.CheckValidity takes
a bare `func() time.Time` rather than this interface, and profile_printer.go:57
converts between the two with a method value. go-xcode has two clock conventions;
this commit unifies the interface half of it.
