+++ timeutil/time_provider.go
@@ -1,0 +1,16 @@
+package timeutil
+
+import "time"
+
+// TimeProvider ...
+type TimeProvider interface {
+	Now() time.Time
+}
+
+// DefaultTimeProvider ...
+type DefaultTimeProvider struct{}
+
+// Now ...
+func (DefaultTimeProvider) Now() time.Time {
+	return time.Now()
+}
